New plugins.wgts.common package; move CNV code to common.cnv; simplify import path in process_CNA_data.r
