+++ src/lib/djerba/plugins/cnv/cnv_template.html
@@ -2,7 +2,7 @@
 ## This file is the HTML code for the Copy Number Variation (CNV) section of the report.
 
 <%
-  import djerba.plugins.cnv.constants as cnv
+  import djerba.plugins.wgts.common.cnv.constants as cnv
   from djerba.plugins.cnv.html import make_table_header, make_table_rows
   from djerba.util.html import html_builder
 %>
