+++ src/lib/djerba/plugins/wgts/common/cnv/constants.py
